requirements.txt
llama-index==0.12.48
llama-index-llms-openai==0.4.7
llama-index-embeddings-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-vector-stores-qdrant==0.6.1
qdrant-client==1.14.3
openai==1.95.1
pypdf>=5.1.0
tqdm==4.66.1
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
nest-asyncio==1.6.0